restructure layout
